Add support for Binlog Transaction Compression
This patch adds support for Binlog Transaction Compression, which has
been available since MySQL 8.0.20. It is disabled by default, but it is
widely used in production.

The patch adds a new event type, TransactionPayloadEvent and its
necessary methods to parse binlog event TRANSACTION_PAYLOAD_EVENT(0x28).

The patch also adds a method to decompress the payload if it has been
compressed with ZSTD.

More information about the feature can be found here:

https://dev.mysql.com/worklog/task/?id=3549
https://github.com/mysql/mysql-server/commit/1e5ae348c6f760b16a3972d84956358606a30041